Fix date today indicator
